Cleaned up some type casting and math that wasn't really needed in the PID CPU usage cals.  I tried to run down why the percentage would sometimes be greater than 100.  I couldn't adequately solve it, but then found that the reason is likely because it is using more than 1 core in those cases so the behavior is normal (since I'm using aggregate data to get the difference in total CPU time). I ran the code for over 14 hours without failure so I would call it adqeuately burned in.  Ready for submission.

+++ include/linux_parser.h
@@ -54,7 +54,7 @@ std::string Command(int pid);
 std::string Ram(int pid);
 std::string Uid(int pid);
 std::string User(int pid);
-float ActiveCpuTime(int pid);
+long ActiveCpuTime(int pid);
 long STime(int pid);
 
 // Helper Functions
